Expose and document `get_location` function
